[pack #179] refactor bundle
